Sphinx: use simplified onion proposed in PR145 address and hop-payload fields have fused, and we are now routing based on channel short ids
