+++ eclair-node/src/test/scala/fr/acinq/eclair/crypto/SphinxSpec.scala
@@ -53,65 +53,39 @@ class SphinxSpec extends FunSuite {
 
   /*
   filler = 0x80b23adf2c3947dab3b92bb1c5f70fa32f866cc09aff45c5bc4da1ed291660791aa9a1c5d28fbf8f4ecb4bf7c0a9454c82923e24c7fd0d192ea8e2ead1f17388341f313915949b602d5de1f5cb94b77d42c7dfe84edc13cf9acf541e8243989d967e7d568f26f9209bb52d9f90bfa902f3ec6e3ae9d6215c432206cd3132b69257408002aa020f2fbae32a2d5abee0a3c9fd56329b033939cd6366fbb339aa14
-  hop_filler = 0x2e86897a3ae52daba4a5940cfc305ae15e9a0f8a8ac1033a15d8a14819acab6503c9df44cdaaf30629283e3458844a44a5c4bfdebdcb15fd3edb8e286124d7b47fa7a56bcc5655d2ad9809f108f238e5
    */
   test("generate filler") {
     val (_, sharedsecrets) = computeEphemerealPublicKeysAndSharedSecrets(sessionKey, publicKeys)
     val filler = generateFiller("rho", sharedsecrets.dropRight(1), 40, 20)
     assert(filler == BinaryData("0x80b23adf2c3947dab3b92bb1c5f70fa32f866cc09aff45c5bc4da1ed291660791aa9a1c5d28fbf8f4ecb4bf7c0a9454c82923e24c7fd0d192ea8e2ead1f17388341f313915949b602d5de1f5cb94b77d42c7dfe84edc13cf9acf541e8243989d967e7d568f26f9209bb52d9f90bfa902f3ec6e3ae9d6215c432206cd3132b69257408002aa020f2fbae32a2d5abee0a3c9fd56329b033939cd6366fbb339aa14"))
-
-    val hopFiller = generateFiller("gamma", sharedsecrets.dropRight(1), 20, 20)
-    assert(hopFiller == BinaryData("0x2e86897a3ae52daba4a5940cfc305ae15e9a0f8a8ac1033a15d8a14819acab6503c9df44cdaaf30629283e3458844a44a5c4bfdebdcb15fd3edb8e286124d7b47fa7a56bcc5655d2ad9809f108f238e5"))
   }
 
-  test("parse packet") {
-    val packet: BinaryData = "0x0102eec7245d6b7d2ccb30380bfbe2a3648cd7a942653f5aa340edcea1f2836866195fcf66568cad9ad9d61c8b05ffeb2ef00bf32dbff12a026817cd410d83bd65c4fc27db8a75033c527e678028eb2b95f58d04baa9986e83030fb5577e0543f62394cca6d995cf368aface565d15e778db79b7ff43a21abbd556d1e1b33753959a62e6bcb5220cda916d603c2702554c4dc17a8896af5d54c7815efcd093a0f6bad14a4d7622df88ee7cd1fe512882beb43b81f7cf3e7e633262538a7eca5f76f8434bd2215c7dda800d49ef34caf74bb4f1064f72e28fa39de96ff993cf51e26512faa0e98fa387f517c3bc4a65f6d8ca27af1d81025a85edf9e3ea7f580854c5d3ce537b955ff216c7a88dfc0bb795bfdf053300c70a1f6f23126c83b3c96a806e0cda3c3bb5cf1c57e77e25aafd117f2c559c914797cdd27440d7d033ef987178e337a899b597a34cd6f471ce40ca44b234fa0730603b1d4ad215bce719d02b98882268551912ce1f14a3a8c968649b09fb9acf69f48046e2464536d734f705e2d4a996c176786ec437b5d0f62731b251c56f67ae4169167fab24673af60d88d3252bf38f8c1ebd85986bbe25f7f0a4399d947d5b15b7ad9050b3a4f67b0720e372a56fa2da6468ec436ca7425bfedccb4186dbbc28663941a2b5473ff853dbfce8327ffb1209940d5b3c9f8d6643d11238da66e35715b5d5410b93bb5c9ed25e7695c3d2a8073e7e5373dcbdab46c50aba3d238573b333b68bdf1c209933e75e27960fc78880e8429a88c24878c2abd541a0afd0e0c364517e0e8064a94ff1cde8885b19d60d30304c5f7bbc7afa64befbdd2095b47730fdcc6aacd6cf927b6c981efad29de9c575c8663b545dc57e7dbc1b7a09bf6584d5c5b1e748c0300903d09b8b33adf548b0f4a9198c7b0f5c4b4c0e3af9fcc9af4860a23aec993996e36aefff3157259147f0536ea64d23329178f3dd95e37e0019e5ab4654325b59544bf3caa891617b59978b3d21785d36166721358a99c90bd2b8364c475b7f8058219269e561303dfd1d84042543e85d954bcd5eeb4ff897aeb2988ed3d905b3c73377be03ef8817d8595f2596e7afa6a0aa121b176738fd578be920b9cf778bd92808e5bd3d8d0decd00bdc2560a22921f138af042f38c923693494237844c13c85b4d0235c462af39d519d4fbb0309805fff51a51d2a1675eb1be407e116535f455d966fc5a33d05d1b7b7aa4cee07f1bfff801a9bd08a9c6e730771a4bf6a646c2f9e2a90bda866a6fecb3e79981f0048309aac743972adda805075163ed26a5af81bc8fe32606fc23d362dd240c5b601b78cfb31fc350de0cacc356fd62dbf6d6fcdaed073647490573c6eca5432b566cde10369f984b036310991b8965d71c85a667ee5b1d2d043aa4e4d239fda39474626a7840708cb9b7a5311d6f187f8f485ec91e5768007db224e38ff189dd809b1604cf067fa3fd54fee9c9176396cef8664557e2550f89858071da37a96c17c2a5c1e73feeab9b09d32ba7efbc6b18f9b1c4d85abb2dfb4648e0e6b12a3dd9fe96e6e5317ffe32c7b4b3796e8599b7a4caf4ad1b980350b8e7c8544871d7702f3e8c0a265bd658389fde9850570877856e66e29e16bdd9f2a6f6106e454e531d8b7062d3086b8c61f88677bfdb5e0b8bb6654e206e2123edb48d2b18c4316ab393ae41e570e144af52a7e5d3515260b387750e205596828940376ae31e55476b7fa3d5d4abf2ab586ece8acf42c4b2ca0c4"
-
-    val ParsedPacket(payload0, address0, packet1, sharedSecret0) = parsePacket(privKeys(0), associatedData, packet)
-    val ParsedPacket(payload1, address1, packet2, sharedSecret1) = parsePacket(privKeys(1), associatedData, packet1)
-    val ParsedPacket(payload2, address2, packet3, sharedSecret2) = parsePacket(privKeys(2), associatedData, packet2)
-    val ParsedPacket(payload3, address3, packet4, sharedSecret3) = parsePacket(privKeys(3), associatedData, packet3)
-    val ParsedPacket(payload4, address4, packet5, sharedSecret4) = parsePacket(privKeys(4), associatedData, packet4)
-
-    assert(Seq(payload0, payload1, payload2, payload3, payload4) == payloads)
-    assert(Seq(address0, address1, address2, address3, address4) == Seq(publicKeys(1).hash160, publicKeys(2).hash160, publicKeys(3).hash160, publicKeys(4).hash160, zeroes(20)))
-
-    val (_, sharedsecrets) = computeEphemerealPublicKeysAndSharedSecrets(sessionKey, publicKeys)
-    assert(Seq(sharedSecret0, sharedSecret1, sharedSecret2, sharedSecret3, sharedSecret4) == sharedsecrets)
+  test("parse 1-hop packet") {
+    val raw = BinaryData("0102eec7245d6b7d2ccb30380bfbe2a3648cd7a942653f5aa340edcea1f2836866194c4fbecf0a2a21892c45d9b5aa509a365d7d022ca4b0021183e62e8373f4a52c035fd754061ab9e0bc887beff8c95fdb878f7b3a7141453e5f8d22b6351810ae541ce499a09b4a9d9f80d1845c8960c85fc6d1a87bd24b2ce49922898e9353fa268086c00ae8b7f718405b72ad380cdbb38c85e02a00427eb4bdbda8fcd42b44708a9efde49cf753b75ebb389bf84d0bfbf58590e510e034572a01e409c30939e2e4a090ecc89c371820af54e06e4ad5495d4e58718385cca5414552e078fedf284fdc2cc5c070cba21a6a8d4b77525ddbc9a9fca9b2f29aac5783ee8badd709f81c73ff60556cf2ee623af073b5a84799acc1ca46b764f74b97068c7826cc0579794a540d7a55e49eac26a6930340132e946a983240b0cd1b732e305c1042f590c4b26f140fc1cab3ee6f620958e0979f85eddf586c410ce42e93a4d7c803ead45fc47cf4396d284632314d789e73cf3f534126c63fe244069d9e8a7c4f98e7e530fc588e648ef4e641364981b5377542d5e7a4aaab6d35f6df7d3a9d7ca715213599ee02c4dbea4dc78860febe1d29259c64b59b3333ffdaebbaff4e7b31c27a3791f6bf848a58df7c69bb2b1852d2ad357b9919ffdae570b27dc709fba087273d3a4de9e6a6be66db647fb6a8d1a503b3f481befb96745abf5cc4a6bba0f780d5c7759b9e303a2a6b17eb05b6e660f4c474959db183e1cae060e1639227ee0bca03978a238dc4352ed764da7d4f3ed5337f6d0376dff72615beeeeaaeef79ab93e4bcbf18cd8424eb2b6ad7f33d2b4ffd5ea08372e6ed1d984152df17e04c6f73540988d7dd979e020424a163c271151a255966be7edef42167b8facca633649739bab97572b485658cde409e5d4a0f653f1a5911141634e3d2b6079b19347df66f9820755fd517092dae62fb278b0bafcc7ad682f7921b3a455e0c6369988779e26f0458b31bffd7e4e5bfb31944e80f100b2553c3b616e75be18328dc430f6618d55cd7d0962bb916d26ed4b117c46fa29e0a112c02c36020b34a96762db628fa3490828ec2079962ad816ef20ea0bca78fb2b7f7aedd4c47e375e64294d151ff03083730336dea64934003a27730cc1c7dec5049ddba8188123dd191aa71390d43a49fb792a3da7082efa6cced73f00eccea18145fbc84925349f7b552314ab8ed4c491e392aed3b1f03eb79474c294b42e2eba1528da26450aa592cba7ea22e965c54dff0fd6fdfd6b52b9a0f5f762e27fb0e6c3cd326a1ca1c5973de9be881439f702830affeb0c034c18ac8d5c2f135c964bf69de50d6e99bde88e90321ba843d9753c8f83666105d25fafb1a11ea22d62ef6f1fc34ca4e60c35d69773a104d9a44728c08c20b6314327301a2c400a71e1424c12628cf9f4a67990ade8a2203b0edb96c6082d4673b7309cd52c4b32b02951db2f66c6c72bd6c7eac2b50b83830c75cdfc3d6e9c2b592c45ed5fa5f6ec0da85710b7e1562aea363e28665835791dc574d9a70b2e5e2b9973ab590d45b94d244fc")
+    val Sphinx.ParsedPacket(payload, nextPacket, sharedSecret) = Sphinx.parsePacket(privKeys(0), associatedData, raw)
+    assert(nextPacket == BinaryData("01028f9438bfbf7feac2e108d677e3a82da596be706cc1cf342b75c7b7e22bf4e6e2000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004256926c5a55b01cd0aca21fe5f9c907691fb026d0c56788b03ca3f08db0abb9f901098dde2ec4003568bc3ca27475ff86a7cb0aab"))
   }
 
-  test("generate last packet") {
-    val (ephemerealPublicKeys, sharedsecrets) = computeEphemerealPublicKeysAndSharedSecrets(sessionKey, publicKeys)
-    val filler = generateFiller("rho", sharedsecrets.dropRight(1), 40, 20)
-    val hopFiller = generateFiller("gamma", sharedsecrets.dropRight(1), 20, 20)
-
-    // build the last packet and apply obfuscation in reverse node order
-    val nextPacket = makeNextPacket(LAST_ADDRESS, payloads(4), associatedData, ephemerealPublicKeys(4), sharedsecrets(4), LAST_PACKET, filler, hopFiller)
-    val nextPacket1 = makeNextPacket(publicKeys(4).hash160, payloads(3), associatedData, ephemerealPublicKeys(3), sharedsecrets(3), nextPacket)
-    val nextPacket2 = makeNextPacket(publicKeys(3).hash160, payloads(2), associatedData, ephemerealPublicKeys(2), sharedsecrets(2), nextPacket1)
-    val nextPacket3 = makeNextPacket(publicKeys(2).hash160, payloads(1), associatedData, ephemerealPublicKeys(1), sharedsecrets(1), nextPacket2)
-    val nextPacket4 = makeNextPacket(publicKeys(1).hash160, payloads(0), associatedData, ephemerealPublicKeys(0), sharedsecrets(0), nextPacket3)
-    assert(nextPacket4 == BinaryData("0x0102eec7245d6b7d2ccb30380bfbe2a3648cd7a942653f5aa340edcea1f2836866195fcf66568cad9ad9d61c8b05ffeb2ef00bf32dbff12a026817cd410d83bd65c4fc27db8a75033c527e678028eb2b95f58d04baa9986e83030fb5577e0543f62394cca6d995cf368aface565d15e778db79b7ff43a21abbd556d1e1b33753959a62e6bcb5220cda916d603c2702554c4dc17a8896af5d54c7815efcd093a0f6bad14a4d7622df88ee7cd1fe512882beb43b81f7cf3e7e633262538a7eca5f76f8434bd2215c7dda800d49ef34caf74bb4f1064f72e28fa39de96ff993cf51e26512faa0e98fa387f517c3bc4a65f6d8ca27af1d81025a85edf9e3ea7f580854c5d3ce537b955ff216c7a88dfc0bb795bfdf053300c70a1f6f23126c83b3c96a806e0cda3c3bb5cf1c57e77e25aafd117f2c559c914797cdd27440d7d033ef987178e337a899b597a34cd6f471ce40ca44b234fa0730603b1d4ad215bce719d02b98882268551912ce1f14a3a8c968649b09fb9acf69f48046e2464536d734f705e2d4a996c176786ec437b5d0f62731b251c56f67ae4169167fab24673af60d88d3252bf38f8c1ebd85986bbe25f7f0a4399d947d5b15b7ad9050b3a4f67b0720e372a56fa2da6468ec436ca7425bfedccb4186dbbc28663941a2b5473ff853dbfce8327ffb1209940d5b3c9f8d6643d11238da66e35715b5d5410b93bb5c9ed25e7695c3d2a8073e7e5373dcbdab46c50aba3d238573b333b68bdf1c209933e75e27960fc78880e8429a88c24878c2abd541a0afd0e0c364517e0e8064a94ff1cde8885b19d60d30304c5f7bbc7afa64befbdd2095b47730fdcc6aacd6cf927b6c981efad29de9c575c8663b545dc57e7dbc1b7a09bf6584d5c5b1e748c0300903d09b8b33adf548b0f4a9198c7b0f5c4b4c0e3af9fcc9af4860a23aec993996e36aefff3157259147f0536ea64d23329178f3dd95e37e0019e5ab4654325b59544bf3caa891617b59978b3d21785d36166721358a99c90bd2b8364c475b7f8058219269e561303dfd1d84042543e85d954bcd5eeb4ff897aeb2988ed3d905b3c73377be03ef8817d8595f2596e7afa6a0aa121b176738fd578be920b9cf778bd92808e5bd3d8d0decd00bdc2560a22921f138af042f38c923693494237844c13c85b4d0235c462af39d519d4fbb0309805fff51a51d2a1675eb1be407e116535f455d966fc5a33d05d1b7b7aa4cee07f1bfff801a9bd08a9c6e730771a4bf6a646c2f9e2a90bda866a6fecb3e79981f0048309aac743972adda805075163ed26a5af81bc8fe32606fc23d362dd240c5b601b78cfb31fc350de0cacc356fd62dbf6d6fcdaed073647490573c6eca5432b566cde10369f984b036310991b8965d71c85a667ee5b1d2d043aa4e4d239fda39474626a7840708cb9b7a5311d6f187f8f485ec91e5768007db224e38ff189dd809b1604cf067fa3fd54fee9c9176396cef8664557e2550f89858071da37a96c17c2a5c1e73feeab9b09d32ba7efbc6b18f9b1c4d85abb2dfb4648e0e6b12a3dd9fe96e6e5317ffe32c7b4b3796e8599b7a4caf4ad1b980350b8e7c8544871d7702f3e8c0a265bd658389fde9850570877856e66e29e16bdd9f2a6f6106e454e531d8b7062d3086b8c61f88677bfdb5e0b8bb6654e206e2123edb48d2b18c4316ab393ae41e570e144af52a7e5d3515260b387750e205596828940376ae31e55476b7fa3d5d4abf2ab586ece8acf42c4b2ca0c4"))
-
-    // same as above but in one single step
-    // this is what clients will use
-    val OnionPacket(packet, _) = makePacket(sessionKey, publicKeys, payloads, associatedData)
-    assert(packet == nextPacket4)
+  test("parse 2-hop packet") {
+    val raw = BinaryData("0102eec7245d6b7d2ccb30380bfbe2a3648cd7a942653f5aa340edcea1f283686619a84f87d4cf02e4f0387456161c04017b7e5e3a72a4b0021183e62e8373f4a52c035fd754061ab9e0bc887beff8c95fdb878f7b3a71d07f2d00f9c77994251e861f162dde3da7bda32dc2d59599fb5b04a0dacb131d6b2f9e0c766dca91fc29ec8bf8e1c482940a72a5d3e5eeefdf9910c08a542e8c43f8111c9aed119dc3cf18e1abe4db1802a58b8e177752763920dca5e06ff30e3337ba365c0972bf5efb526ad11a3baf39843c89331ffb6a4d0ee8f0b95c59b5a33e3e0464ab184f10860bb582c6a5760ec04680f6b1bbde967158fda50d1f580c36634a5d7abe0185f776b89bce245e8777fb2b843979406056c302ccffd69a88ce6b2a52dc6f32272e25d91bf72f055667873fa9a7aaec4eb90d432ed1c40a921feb844886f413f4083f1b7de79bd72e236e671cc623df4b8a5c5cebe2656d1cf8d43685d8ceb3a2c0c99f69265c887e618633e00e172a38db887c8c4b368c681a154afcdf399d755c442babcf03d1ab7aa3d28622639ea9596e7fa7774348b9880d02459819633c5669d80b58a8c3761812930887d88a17fe57234242dc592720eb17d408c0db18b164613bc0c0f77cf36e8bd27a573d7f204d2eb6cf4a72675499a4c426b3c9c484f1c24f007bbe91b25b1d0133b912deb9b9c2f5fa087dcec26d9c4624c9575afd80ff6be481c55d864c4a85801adf2a499e0a51846d954eabcd50de52b0248360e3a54fbb4f7ced9cb7cbb5a987003c622fb269d1517eddc9142c3ffa87e0973173eff9d5f24105ca0b6d50acdef393c8ee70f6d221853e2fc265a9a2792cc7b61f26823ec2d3db1bd2287c5d665aa3143e1f1216fef8f7236801e860925567bdb2868ef80bdc9b5b83b1cdbf17c4cc740bbec1db724e9cb5ea2b33442e5ac1624236c3be33dc8c9d7bd5415bfeb72f4f44fbf9a88bd63e585ff5572dca1cce859defa6dcc115f1ec71d38cf1547e6dc07ffb0dc156b81acadd91a72386a4b809f05999ab277204e115f7ca54245774338ce027f3e11f24561412e742e64854d6df45a1dc109b0bfa0ac7bfb70ffb93801099f6262e807d1cd8526f64a6683a3d624fcbd5dfdc23599e9015f84926bc66fa29c9003308e58500da4cd52b89f4d287b1e03838b13ce9d9b16a3c3e560937dcb3aa7e02b6b628297ca4ec4ee5181dd987908a9f38966c8eece3024476273f9ab222825ed1523ad6624c213ab83124590a32ad4abb63b2bace9563c559cf7ab5d5320ca199890265c9f0495b4714f24438c16e12b55bae6059aeab65ef87dec58ebfcfb81ca8413f4fe3dff01037f48e68b7d5bc745f4873e054af9645e40b9dbc042b312c5949e876faf22da920a546043644d688356f185b624df2908e344765ea0e4036db041532983496ec34a0865150d71319c89bc86a1ff216f87ce478dd6b264990eceea7eb135eede449d88afc87df0c64dbf40bbdaa817ce0fd5a7eb7b0b0053b9cc01f2a16b63ef5bfb8baf1c6c3feada8a0866f197e60bb43eac5b59fc9d55adcda2e")
+    val Sphinx.ParsedPacket(payload, nextPacket, sharedSecret) = Sphinx.parsePacket(privKeys(0), associatedData, raw)
+    val Sphinx.ParsedPacket(payload1, nextPacket1, sharedSecret1) = Sphinx.parsePacket(privKeys(1), associatedData, nextPacket)
+    assert(Seq(payload, payload1) == payloads.take(2))
+    val Sphinx.Packet(_, _, hmac, _) = Sphinx.Packet.read(nextPacket1)
+    assert(hmac == Sphinx.zeroes(Sphinx.MacLength))
   }
 
-  test("generate return messages") {
-    val failure = TemporaryChannelFailure
-    val (ephkeys, sharedsecrets) = computeEphemerealPublicKeysAndSharedSecrets(sessionKey, publicKeys)
+  test("create 1-hop packet") {
+    val Sphinx.OnionPacket(onion, _) = Sphinx.makePacket(sessionKey, publicKeys.take(1), payloads.take(1), associatedData)
+    val Sphinx.ParsedPacket(payload, nextPacket, sharedSecret) = Sphinx.parsePacket(privKeys(0), associatedData, onion)
+    assert(onion == BinaryData("0102eec7245d6b7d2ccb30380bfbe2a3648cd7a942653f5aa340edcea1f2836866194c4fbecf0a2a21892c45d9b5aa509a365d7d022ca4b0021183e62e8373f4a52c035fd754061ab9e0bc887beff8c95fdb878f7b3a7141453e5f8d22b6351810ae541ce499a09b4a9d9f80d1845c8960c85fc6d1a87bd24b2ce49922898e9353fa268086c00ae8b7f718405b72ad380cdbb38c85e02a00427eb4bdbda8fcd42b44708a9efde49cf753b75ebb389bf84d0bfbf58590e510e034572a01e409c30939e2e4a090ecc89c371820af54e06e4ad5495d4e58718385cca5414552e078fedf284fdc2cc5c070cba21a6a8d4b77525ddbc9a9fca9b2f29aac5783ee8badd709f81c73ff60556cf2ee623af073b5a84799acc1ca46b764f74b97068c7826cc0579794a540d7a55e49eac26a6930340132e946a983240b0cd1b732e305c1042f590c4b26f140fc1cab3ee6f620958e0979f85eddf586c410ce42e93a4d7c803ead45fc47cf4396d284632314d789e73cf3f534126c63fe244069d9e8a7c4f98e7e530fc588e648ef4e641364981b5377542d5e7a4aaab6d35f6df7d3a9d7ca715213599ee02c4dbea4dc78860febe1d29259c64b59b3333ffdaebbaff4e7b31c27a3791f6bf848a58df7c69bb2b1852d2ad357b9919ffdae570b27dc709fba087273d3a4de9e6a6be66db647fb6a8d1a503b3f481befb96745abf5cc4a6bba0f780d5c7759b9e303a2a6b17eb05b6e660f4c474959db183e1cae060e1639227ee0bca03978a238dc4352ed764da7d4f3ed5337f6d0376dff72615beeeeaaeef79ab93e4bcbf18cd8424eb2b6ad7f33d2b4ffd5ea08372e6ed1d984152df17e04c6f73540988d7dd979e020424a163c271151a255966be7edef42167b8facca633649739bab97572b485658cde409e5d4a0f653f1a5911141634e3d2b6079b19347df66f9820755fd517092dae62fb278b0bafcc7ad682f7921b3a455e0c6369988779e26f0458b31bffd7e4e5bfb31944e80f100b2553c3b616e75be18328dc430f6618d55cd7d0962bb916d26ed4b117c46fa29e0a112c02c36020b34a96762db628fa3490828ec2079962ad816ef20ea0bca78fb2b7f7aedd4c47e375e64294d151ff03083730336dea64934003a27730cc1c7dec5049ddba8188123dd191aa71390d43a49fb792a3da7082efa6cced73f00eccea18145fbc84925349f7b552314ab8ed4c491e392aed3b1f03eb79474c294b42e2eba1528da26450aa592cba7ea22e965c54dff0fd6fdfd6b52b9a0f5f762e27fb0e6c3cd326a1ca1c5973de9be881439f702830affeb0c034c18ac8d5c2f135c964bf69de50d6e99bde88e90321ba843d9753c8f83666105d25fafb1a11ea22d62ef6f1fc34ca4e60c35d69773a104d9a44728c08c20b6314327301a2c400a71e1424c12628cf9f4a67990ade8a2203b0edb96c6082d4673b7309cd52c4b32b02951db2f66c6c72bd6c7eac2b50b83830c75cdfc3d6e9c2b592c45ed5fa5f6ec0da85710b7e1562aea363e28665835791dc574d9a70b2e5e2b9973ab590d45b94d244fc"))
+  }
 
-    // error packet created by the last node
-    val error0 = createErrorPacket(sharedsecrets.last, failure)
-
-    // error packet received by the origin node
-    val error = sharedsecrets.dropRight(1).reverse.foldLeft(error0)(forwardErrorPacket)
-
-    val Some(ErrorPacket(pubkey, failure1)) = parseErrorPacket(error, sharedsecrets.zip(publicKeys))
-    assert(pubkey == publicKeys.last)
-    assert(failure1 == failure)
+  test("create 2-hop packet") {
+    val Sphinx.OnionPacket(onion, _) = Sphinx.makePacket(sessionKey, publicKeys.take(2), payloads.take(2), associatedData)
+    val Sphinx.ParsedPacket(payload, nextPacket, sharedSecret) = Sphinx.parsePacket(privKeys(0), associatedData, onion)
+    val Sphinx.ParsedPacket(payload1, nextPacket1, sharedSecret1) = Sphinx.parsePacket(privKeys(1), associatedData, nextPacket)
+    assert(onion == BinaryData("0102eec7245d6b7d2ccb30380bfbe2a3648cd7a942653f5aa340edcea1f283686619a84f87d4cf02e4f0387456161c04017b7e5e3a72a4b0021183e62e8373f4a52c035fd754061ab9e0bc887beff8c95fdb878f7b3a71d07f2d00f9c77994251e861f162dde3da7bda32dc2d59599fb5b04a0dacb131d6b2f9e0c766dca91fc29ec8bf8e1c482940a72a5d3e5eeefdf9910c08a542e8c43f8111c9aed119dc3cf18e1abe4db1802a58b8e177752763920dca5e06ff30e3337ba365c0972bf5efb526ad11a3baf39843c89331ffb6a4d0ee8f0b95c59b5a33e3e0464ab184f10860bb582c6a5760ec04680f6b1bbde967158fda50d1f580c36634a5d7abe0185f776b89bce245e8777fb2b843979406056c302ccffd69a88ce6b2a52dc6f32272e25d91bf72f055667873fa9a7aaec4eb90d432ed1c40a921feb844886f413f4083f1b7de79bd72e236e671cc623df4b8a5c5cebe2656d1cf8d43685d8ceb3a2c0c99f69265c887e618633e00e172a38db887c8c4b368c681a154afcdf399d755c442babcf03d1ab7aa3d28622639ea9596e7fa7774348b9880d02459819633c5669d80b58a8c3761812930887d88a17fe57234242dc592720eb17d408c0db18b164613bc0c0f77cf36e8bd27a573d7f204d2eb6cf4a72675499a4c426b3c9c484f1c24f007bbe91b25b1d0133b912deb9b9c2f5fa087dcec26d9c4624c9575afd80ff6be481c55d864c4a85801adf2a499e0a51846d954eabcd50de52b0248360e3a54fbb4f7ced9cb7cbb5a987003c622fb269d1517eddc9142c3ffa87e0973173eff9d5f24105ca0b6d50acdef393c8ee70f6d221853e2fc265a9a2792cc7b61f26823ec2d3db1bd2287c5d665aa3143e1f1216fef8f7236801e860925567bdb2868ef80bdc9b5b83b1cdbf17c4cc740bbec1db724e9cb5ea2b33442e5ac1624236c3be33dc8c9d7bd5415bfeb72f4f44fbf9a88bd63e585ff5572dca1cce859defa6dcc115f1ec71d38cf1547e6dc07ffb0dc156b81acadd91a72386a4b809f05999ab277204e115f7ca54245774338ce027f3e11f24561412e742e64854d6df45a1dc109b0bfa0ac7bfb70ffb93801099f6262e807d1cd8526f64a6683a3d624fcbd5dfdc23599e9015f84926bc66fa29c9003308e58500da4cd52b89f4d287b1e03838b13ce9d9b16a3c3e560937dcb3aa7e02b6b628297ca4ec4ee5181dd987908a9f38966c8eece3024476273f9ab222825ed1523ad6624c213ab83124590a32ad4abb63b2bace9563c559cf7ab5d5320ca199890265c9f0495b4714f24438c16e12b55bae6059aeab65ef87dec58ebfcfb81ca8413f4fe3dff01037f48e68b7d5bc745f4873e054af9645e40b9dbc042b312c5949e876faf22da920a546043644d688356f185b624df2908e344765ea0e4036db041532983496ec34a0865150d71319c89bc86a1ff216f87ce478dd6b264990eceea7eb135eede449d88afc87df0c64dbf40bbdaa817ce0fd5a7eb7b0b0053b9cc01f2a16b63ef5bfb8baf1c6c3feada8a0866f197e60bb43eac5b59fc9d55adcda2e"))
   }
 
   test("last node replies with an error message") {
@@ -122,16 +96,16 @@ class SphinxSpec extends FunSuite {
 
     // each node parses and forwards the packet
     // node #0
-    val ParsedPacket(payload0, address0, packet1, sharedSecret0) = parsePacket(privKeys(0), associatedData, packet)
+    val ParsedPacket(payload0, packet1, sharedSecret0) = parsePacket(privKeys(0), associatedData, packet)
     // node #1
-    val ParsedPacket(payload1, address1, packet2, sharedSecret1) = parsePacket(privKeys(1), associatedData, packet1)
+    val ParsedPacket(payload1, packet2, sharedSecret1) = parsePacket(privKeys(1), associatedData, packet1)
     // node #2
-    val ParsedPacket(payload2, address2, packet3, sharedSecret2) = parsePacket(privKeys(2), associatedData, packet2)
+    val ParsedPacket(payload2, packet3, sharedSecret2) = parsePacket(privKeys(2), associatedData, packet2)
     // node #3
-    val ParsedPacket(payload3, address3, packet4, sharedSecret3) = parsePacket(privKeys(3), associatedData, packet3)
+    val ParsedPacket(payload3, packet4, sharedSecret3) = parsePacket(privKeys(3), associatedData, packet3)
     // node #4
-    val ParsedPacket(payload4, address4, packet5, sharedSecret4) = parsePacket(privKeys(4), associatedData, packet4)
-    assert(address4 == LAST_ADDRESS)
+    val ParsedPacket(payload4, packet5, sharedSecret4) = parsePacket(privKeys(4), associatedData, packet4)
+    assert(Packet.isLastPacket(packet5))
 
     // node #4 want to reply with an error message
     val error = createErrorPacket(sharedSecret4, TemporaryNodeFailure)
@@ -164,11 +138,11 @@ class SphinxSpec extends FunSuite {
 
     // each node parses and forwards the packet
     // node #0
-    val ParsedPacket(payload0, address0, packet1, sharedSecret0) = parsePacket(privKeys(0), associatedData, packet)
+    val ParsedPacket(payload0, packet1, sharedSecret0) = parsePacket(privKeys(0), associatedData, packet)
     // node #1
-    val ParsedPacket(payload1, address1, packet2, sharedSecret1) = parsePacket(privKeys(1), associatedData, packet1)
+    val ParsedPacket(payload1, packet2, sharedSecret1) = parsePacket(privKeys(1), associatedData, packet1)
     // node #2
-    val ParsedPacket(payload2, address2, packet3, sharedSecret2) = parsePacket(privKeys(2), associatedData, packet2)
+    val ParsedPacket(payload2, packet3, sharedSecret2) = parsePacket(privKeys(2), associatedData, packet2)
 
     // node #2 want to reply with an error message
     val error = createErrorPacket(sharedSecret2, InvalidRealm)
@@ -203,11 +177,11 @@ object SphinxSpec {
 
   val sessionKey: PrivateKey = PrivateKey(BinaryData("0x4141414141414141414141414141414141414141414141414141414141414141"), compressed = true)
   val payloads = Seq(
-    BinaryData("0x4141414141414141414141414141414141414141"),
-    BinaryData("0x4141414141414141414141414141414141414141"),
-    BinaryData("0x4141414141414141414141414141414141414141"),
-    BinaryData("0x4141414141414141414141414141414141414141"),
-    BinaryData("0x4141414141414141414141414141414141414141")
-  )
+    BinaryData("0x414141414141414141414141414141414100000000000000000000000000000000"),
+    BinaryData("0x414141414141414141414141414141414100000000000000000000000000000000"),
+    BinaryData("0x414141414141414141414141414141414100000000000000000000000000000000"),
+    BinaryData("0x414141414141414141414141414141414100000000000000000000000000000000"),
+    BinaryData("0x414141414141414141414141414141414100000000000000000000000000000000"))
+
   val associatedData: BinaryData = "0x4242424242424242424242424242424242424242424242424242424242424242"
 }
